First commit: DistanceMatrix calculates distances between Phenomatrix.
Most of the hooks are already set up for use in Ruby; can load DistanceMatrix in the console. If you want to know how, look in test/test_distance_matrix_extn.rb.

Basic tests are written for both classes.

Still to be done: knearest algorithm for DistanceMatrix. Do I need to calculate the actual distance matrix, or can it be done on the fly as it is for the nearest (not knearest) algorithm?

Also, if we want this gem to work inline -- within the server, and not as a workling -- we may need to change preloading. That is, there's no reason for predict_matrix to load the WHOLE matrix if we just want a couple of columns. Only source_matrix needs all hands on deck.

Should probably also write a C++ class to handle scoring for cross-validation. Need to think more about this, as it may actually be easier in Ruby and nearly as fast.