Implemented Naive Bayes and base classifier types. Reshuffled source layout so now a great deal more in the .cpp files. Had no choice about it; without doing so, there'd be circular includes which suck.
Not sure if NaiveBayes works quite yet. It's definitely predicting something, but right now, for phenotype 12, looks like everything from mouse (matrix 3) is above 0.9. That kind of worries me. The regular Classifier works, but it really needs to be made pure virtual.

The Ruby interface for DistanceMatrix now accepts a hash of options for NaiveBayes (or whatever classifier). This is converted into cparams, but right now it's done in the DistanceMatrix constructor and not in cparams.h as it should be. Need to fix that.

Both Phenomatrix and DistanceMatrix were placed under the Fastknn module, since I finally figured out how to do so.

Cross-validation still needs to be written.