Created inherited phenomatrix type which accounts for source species rows.
PhenomatrixBase is the original, more or less, but with some excess stuff removed to make it load a little faster.

Incidentally, Phenomatrix should load just as quickly as PhenomatrixBase if you're doing it for a single species and not a pair.

Updated test cases for Phenomatrix to confirm that the two types are loading correctly from the database.