Major refactoring of Connection. Completed implementation of PhenomatrixPair.
Constructed tests for Connection, which is now expected to utilize only a single pqxx::connection, via singleton. Fastknn now loads the Connection immediately, and so Phenomatrix (and other classes) no longer needs a database string or pointer -- just matrix ids.

Added a "mask stack" to PhenomatrixPair so rows can be removed and added back. This is more like an undo buffer; it doesn't store modifications, it stores copies that are different from the originals. Can push and pop them using push_mask and pop_mask, both of which work in Ruby.

Not likely that DistanceMatrix compiles. DistanceMatrix, as originally conceived, is not long for this world.

Oh, and an extra file was added, ruby_conversions.h. It doesn't seem to work, but I want to keep messing with it. I want my conversions to be implicit so I don't have to keep writing functions like rb_push_mask.